+++ pkg/storage/chunk/aws/s3_storage_client.go
@@ -28,6 +28,7 @@ import (
 	awscommon "github.com/weaveworks/common/aws"
 	"github.com/weaveworks/common/instrument"
 
+	cortex_aws "github.com/cortexproject/cortex/pkg/chunk/aws"
 	cortex_s3 "github.com/cortexproject/cortex/pkg/storage/bucket/s3"
 	"github.com/cortexproject/cortex/pkg/util"
 	"github.com/grafana/dskit/flagext"
@@ -123,6 +124,32 @@ func (cfg *S3Config) Validate() error {
 		return errUnsupportedSignatureVersion
 	}
 	return nil
+}
+
+func (cfg *S3Config) ToCortexS3Config() cortex_aws.S3Config {
+	return cortex_aws.S3Config{
+		S3:               cfg.S3,
+		S3ForcePathStyle: cfg.S3ForcePathStyle,
+		BucketNames:      cfg.BucketNames,
+		Endpoint:         cfg.Endpoint,
+		Region:           cfg.Region,
+		AccessKeyID:      cfg.AccessKeyID,
+		SecretAccessKey:  cfg.SecretAccessKey,
+		Insecure:         cfg.Insecure,
+		SSEEncryption:    cfg.SSEEncryption,
+		HTTPConfig:       cfg.HTTPConfig.ToCortexHTTPConfig(),
+		SignatureVersion: cfg.SignatureVersion,
+		SSEConfig:        cfg.SSEConfig,
+		Inject:           cortex_aws.InjectRequestMiddleware(cfg.Inject),
+	}
+}
+
+func (cfg *HTTPConfig) ToCortexHTTPConfig() cortex_aws.HTTPConfig {
+	return cortex_aws.HTTPConfig{
+		IdleConnTimeout:       cfg.IdleConnTimeout,
+		ResponseHeaderTimeout: cfg.ResponseHeaderTimeout,
+		InsecureSkipVerify:    cfg.InsecureSkipVerify,
+	}
 }
 
 type S3ObjectClient struct {